fix(🖍️): fix regression in ColorType enum
