+++ packages/skia/src/skia/types/Image/index.ts
@@ -1,2 +1,3 @@
 export * from "./Image";
 export * from "./ImageFactory";
+export * from "./ColorType";
